feat: Moved initial school search endpoint to `/school` from `/find-organisation/school`
