--- web/src/Web.App/Views/FindOrganisation/School.cshtml
@@ -1,13 +1,0 @@
-@using Web.App.ViewModels
-@model Web.App.ViewModels.FindSchoolViewModel
-@{
-    ViewData[ViewDataKeys.Title] = PageTitles.FindOrganisationSchool;
-}
-
-@using (Html.BeginForm(FormMethod.Post, true, new { novalidate = "novalidate", role = "search" }))
-{
-    @await Html.PartialAsync("Search/_SchoolSearchTerm", new FindSchoolViewModel
-    {
-        Term = Model.Term
-    })
-}
