+++ web/tests/Web.Integration.Tests/Pages/Schools/WhenViewingHomeAsFederation.cs
@@ -89,14 +89,9 @@ public class WhenViewingHomeAsFederation(SchoolBenchmarkingWebAppClient client) 
 
         page = await Client.Follow(anchor);
 
-        if (facetedSearchFeatureEnabled)
-        {
-            DocumentAssert.AssertPageUrl(page, $"{Paths.FindOrganisation.ToAbsolute()}/school");
-        }
-        else
-        {
-            DocumentAssert.AssertPageUrl(page, $"{Paths.FindOrganisation.ToAbsolute()}?method=school");
-        }
+        DocumentAssert.AssertPageUrl(page, facetedSearchFeatureEnabled
+            ? Paths.SchoolSearch.ToAbsolute()
+            : $"{Paths.FindOrganisation.ToAbsolute()}?method=school");
     }
 
     [Fact]
